requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
requests
pytrends
rapidfuzz
